+++ example/lib/services.dart
@@ -1,5 +1,6 @@
 import "dart:async";
 import "package:flutter_bloc_annotations/flutter_bloc_annotations.dart";
+import "bloc.dart";
 
 class SetService extends InputService<int> {
 	@override
@@ -16,5 +17,27 @@ class PrintService extends OutputService<String> {
 	@override
 	void listen(String inputData) {
 		print("Counter set to: $inputData");
+	}
+}
+
+class MaxService extends BLoCService<TestBLoC> {
+	StreamSubscription<String> _counterSub;
+
+	@override
+	void init(TestBLoC bloc) {
+		_counterSub = bloc.counter.listen((String counter) {
+			if(int.parse(counter) >= 20) {
+				print(
+					"Looks like you hit 20!\n"
+					"Resetting..."
+				);
+				bloc.setCounter.add(0);
+			}
+		});
+	}
+
+	@override
+	void dispose() {
+		_counterSub.cancel();
 	}
 }